database clubs and coaches

See merge request wwykret/jwzp_ww_fs!15

+++ src/main/java/jwzp_ww_fs/app/services/CoachesService.java
@@ -3,88 +3,75 @@ package jwzp_ww_fs.app.services;
 import jwzp_ww_fs.app.Exceptions.ClubHasEventsException;
 import jwzp_ww_fs.app.Exceptions.CoachHasEventsException;
 import jwzp_ww_fs.app.models.Coach;
-import jwzp_ww_fs.app.repositories.CoachesRepository;
+import jwzp_ww_fs.app.repositories.CoachRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Service
 public class CoachesService {
-    CoachesRepository repository;
-
-    Map<Integer, Integer> numberOfEventsForCoaches;
-
+    CoachRepository repository;
     @Autowired
-    public CoachesService(CoachesRepository repository) {
+    public CoachesService(CoachRepository repository) {
         this.repository = repository;
-        numberOfEventsForCoaches = new HashMap<>();
     }
 
-    synchronized boolean addEventForCoach(int coachId) {
-        if (!numberOfEventsForCoaches.containsKey(coachId)) return false;
-
-        numberOfEventsForCoaches.replace(coachId, numberOfEventsForCoaches.get(coachId) + 1);
-        return true;
+    synchronized void addEventForCoach(int coachId) {
+        Coach coach = repository.getById(coachId);
+        coach.addEvent();
+        repository.save(coach);
     }
 
-    synchronized boolean subtractEventFromCoach(int coachId) {
-        if (!numberOfEventsForCoaches.containsKey(coachId)) return false;
-        if (numberOfEventsForCoaches.get(coachId) <= 0) return false;
-
-        numberOfEventsForCoaches.replace(coachId, numberOfEventsForCoaches.get(coachId) - 1);
-        return true;
+    synchronized void subtractEventFromCoach(int coachId) {
+        Coach coach = repository.getById(coachId);
+        coach.subEvent();
+        repository.save(coach);
     }
 
-    public Coach addCoach(Coach coach) {
-        // var id = repository.getNextId();
-        // var out = repository.addCoach(coach);
+    synchronized void deleteAllEvents() {
+        List<Coach> coaches = repository.findAll();
+        for (var coach : coaches) {
+            coach.deleteEvents();
+            repository.save(coach);
+        }
+    }
 
-        // numberOfEventsForCoaches.put(id, 0);
-
-        // return out;
+    synchronized public Coach addCoach(Coach coach) {
         return repository.save(coach);
-        // return repository.save(new Coach(coach.firstName(), coach.lastName(), coach.yearOfBirth()));
     }
 
     public Coach patchCoach(int coachId, Coach coach) {
-        // return repository.patchCoachWithId(coachId, coach);
-        Coach coachToUpdate = repository.getById(coachId);
-        
+        var coachToUpdate = repository.findById(coachId).orElse(null);
+        if (coachToUpdate == null) return null;
+
         coachToUpdate.updateData(coach);
-        
         return repository.save(coachToUpdate);
     }
 
-    public Coach removeCoach(int coachId) /*throws CoachHasEventsException*/ {
-        // if (numberOfEventsForCoaches.get(coachId) > 0) throw new CoachHasEventsException();
-        // numberOfEventsForCoaches.remove(coachId);
+    synchronized public Coach removeCoach(int coachId) throws CoachHasEventsException {
+        Coach coach = repository.findById(coachId).orElse(null);
+        if (coach == null) return null;
+        if (!coach.isEmpty()) throw new CoachHasEventsException();
 
-        // return repository.removeCoachWithId(coachId);
         repository.deleteById(coachId);
-        return null;
+        return coach;
     }
 
     public List<Coach> removeAllCoaches() throws ClubHasEventsException {
-        // for (var key : numberOfEventsForCoaches.keySet())
-        //     if (numberOfEventsForCoaches.get(key) > 0) throw new ClubHasEventsException();
+        var coaches = repository.findAll();
+        for (var coach : coaches)
+            if (!coach.isEmpty()) throw new ClubHasEventsException();
 
-        // numberOfEventsForCoaches = new HashMap<>();
-
-        // return repository.removeAllCoaches();
         repository.deleteAll();
-        return null;
+        return coaches;
     }
 
     public List<Coach> getAllCoaches() {
         return repository.findAll();
-        // return repository.getAllCoaches();
     }
 
     public Coach getCoach(int coachId) {
-        // return repository.getCoach(coachId);
-        return repository.findById(coachId).isPresent() ? repository.findById(coachId).get() : null;
+        return repository.findById(coachId).orElse(null);
     }
 }
